添加ServerError
