+++ novels_search/views/except_blueprint.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from sanic import Blueprint
-from sanic.exceptions import NotFound
-from sanic.response import html
+from sanic.exceptions import NotFound, ServerError
+from sanic.response import html, json
 from jinja2 import Environment, PackageLoader, select_autoescape
 
 except_bp = Blueprint('except_blueprint', url_prefix='except')
@@ -21,3 +21,13 @@ def template(tpl, **kwargs):
 @except_bp.exception(NotFound)
 def ignore_404s(request, exception):
     return template('404.html')
+
+
+@except_bp.exception(ServerError)
+async def test(request, exception):
+    return json(
+        {
+            "exception": "{}".format(exception),
+            "status": exception.status_code
+        },
+        status=exception.status_code)
